ref(metrics): move metrics samples endpoint to own file
